+++ test/kml_writer_test.dart
@@ -9,6 +9,20 @@ import 'package:test/test.dart';
 import 'utils.dart';
 
 void main() {
+  test('write empty kml', () async {
+    final gpx = createMinimalGPX();
+    final xml = await File('test/assets/minimal.kml').readAsString();
+
+    expectXml(KmlWriter().asString(gpx, pretty: true), xml);
+  });
+
+  test('write empty kml with metadata', () async {
+    final gpx = createMinimalMetadataGPX();
+    final xml = await File('test/assets/minimal_with_metadata.kml').readAsString();
+
+    expectXml(KmlWriter().asString(gpx, pretty: true), xml);
+  });
+
   test('write kml with multiply points', () async {
     final gpx = createGPXWithWpt();
     final xml = await File('test/assets/wpt.kml').readAsString();